src/main/java/customerApp/Customer.java
package customerApp;

public class Customer {
    private Long id;
    private String fname;
    private String lname;
    private String email;
    private int age;
    private String gender;
    private String location;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getFname() {
        return fname;
    }

    public void setFname(String fname){
        if(fname == null || fname.length() <= 2)
            throw new InvalidValueException();
        this.fname = fname;
    }

    public String getLname() {
        return lname;
    }

    public void setLname(String lname) {
        if(lname == null || lname.length() <= 2)
            throw new InvalidValueException();
        this.lname = lname;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }

    public String getGender() {
        return gender;
    }

    public void setGender(String gender) {
        this.gender = gender;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }
}
